pdfsys/requirements.txt
pypdf==5.0.0
reportlab==4.2.0
pymupdf==1.24.6
Pillow==10.4.0
